third_party/WebKit/LayoutTests/external/wpt/html/browsers/the-window-object/reuse-across-navigation/srcdoc.html

<!doctype html>
<meta charset=utf-8>
<title>Window#reuse</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
async_test(function() {
  var iframe = document.createElement("iframe");
  document.body.appendChild(iframe);

  // For same origin navigations from the about:blank Document created when the
  // browsing context was created, the Window object should be reused.
  window[0].testAttribute = "Hello World!";
  iframe.srcdoc = "<script>window.parent.postMessage(window.testAttribute, '*');</scr" + "ipt>";

  window.onmessage = this.step_func_done(function(e) {
    assert_equals(e.data, "Hello World!");
  });
}, "Reuse Window object");
</script>
